trash/AI_interviewer-master/requirements.txt
numpy
jupyter
pandas
scipy
matplotlib
scikit-learn
openai
load-dotenv
selenium
tensorflow
openpyxl
nltk
keras
transformers
torch
gensim
flask
pyaudio
pydub